[🚚rename]: 충돌 방지를 위해 useRooms를 useCreateRoom 네이밍 변경
